Add remove method to cache; move cache to new file
